Added MANY security fixes done by Mike Garrett and Gordon Dickens. Thank you! This commit is a vanilla copy of the one at http://code.google.com/p/vexim/source/detail?r=9951e1422d5a83881297fdd99342937e99781675. It seems Avleen had forgot about that patch...

+++ vexim/admingroupcontentdeletesubmit.php
@@ -3,20 +3,26 @@
   include_once dirname(__FILE__) . '/config/authpostmaster.php';
   include_once dirname(__FILE__) . '/config/functions.php';
 
+  # confirm that the user is updating a group they are permitted to change before going further  
+  $query = "SELECT * FROM groups WHERE id='{$_REQUEST['group_id']}' AND domain_id='{$_SESSION['domain_id']}'";
+  $result = $db->query($query);
+  if ($result->numRows()<1) {
+	  header ("Location: admingroupchange.php?group_id={$_REQUEST['group_id']}&group_failupdated={$_REQUEST['localpart']}"); 
+	  die();  
+  }
+  
   # validate user_id and group_id
   if (!isset($_REQUEST['member_id']) or !isset($_REQUEST['group_id'])) {
     header("Location: admingroup.php?badname={$_REQUEST_['member_id']}");
     die;
   }
   $query = "DELETE FROM group_contents 
-    WHERE group_id={$_REQUEST['group_id']}
-    AND member_id={$_REQUEST['member_id']}";
+    WHERE group_id='{$_REQUEST['group_id']}'
+    AND member_id='{$_REQUEST['member_id']}'";
   $result = $db->query($query);
   if (!DB::isError($result)) { 
-    header ("Location: admingroupchange.php?group_id={$_REQUEST['group_id']}
-      &group_updated={$_REQUEST['localpart']}"); 
+    header ("Location: admingroupchange.php?group_id={$_REQUEST['group_id']}&group_updated={$_REQUEST['localpart']}"); 
   } else { 
-    header ("Location: adminalias.php?
-      group_failupdated={$_REQUEST['localpart']}"); 
+    header ("Location: admingroupchange.php?group_failupdated={$_REQUEST['localpart']}"); 
   }
 ?>
