+++ vexim/admingroupdelete.php
@@ -1,28 +1,35 @@
 <?php
   include_once dirname(__FILE__) . '/config/variables.php';
   include_once dirname(__FILE__) . '/config/authpostmaster.php';
+  include_once dirname(__FILE__) . "/config/functions.php";
 
   if ($_GET['confirm'] == '1') {
+    # confirm that the user is deleting a group they are permitted to change before going further  
+	$query = "SELECT * FROM groups WHERE id='{$_GET['group_id']}' AND domain_id='{$_SESSION['domain_id']}'";
+	$result = $db->query($query);
+	if ($result->numRows()<1) {
+	  header ("Location: admingroup.php?group_faildeleted={$_GET['localpart']}");
+	  die();  
+	}
+	
     # delete group member first
-    $query = "DELETE FROM group_contents WHERE group_id={$_GET['group_id']}";
+    $query = "DELETE FROM group_contents WHERE group_id='{$_GET['group_id']}'";
     $result = $db->query($query);
     if (!DB::isError($result)) {
       # delete group
       $query = "DELETE FROM groups
-        WHERE id={$_GET['group_id']}
-        AND domain_id={$_SESSION['domain_id']}";
+        WHERE id='{$_GET['group_id']}'
+        AND domain_id='{$_SESSION['domain_id']}'";
       $result = $db->query($query);
       if (!DB::isError($result)) {
         header ("Location: admingroup.php?group_deleted={$_GET['localpart']}");
         die;
       } else {
-        header ("Location: admingroup.php?
-          group_faildeleted={$_GET['localpart']}");
+        header ("Location: admingroup.php?group_faildeleted={$_GET['localpart']}");
         die;
       }
     } else {
-      header ("Location: admingroup.php?
-        group_faildeleted={$_GET['localpart']}");
+      header ("Location: admingroup.php?group_faildeleted={$_GET['localpart']}");
       die;
     }
   } else if ($_GET['confirm'] == 'cancel') {
