+++ vexim/adminuserdelete.php
@@ -4,14 +4,37 @@
   include_once dirname(__FILE__) . '/config/functions.php';
   include_once dirname(__FILE__) . '/config/httpheaders.php';
 
+# confirm that the postmaster is looking to delete a user they are permitted to change before going further  
+$query = "SELECT * FROM users WHERE user_id='{$_GET['user_id']}'
+	AND domain_id='{$_SESSION['domain_id']}'	
+	AND (type='local' OR type='piped')";
+$result = $db->query($query);
+if ($result->numRows()<1) {
+  header ("Location: adminuser.php?faildeleted={$_GET['localpart']}");
+  die();  
+}
+
+  
 if ($_GET['confirm'] == '1') {
 
+  # prevent deleting the last admin
+  $query = "SELECT user_id AS count FROM users 
+    WHERE admin=1 AND domain_id='{$_SESSION['domain_id']}'
+	AND (type='local' OR type='piped')
+    AND user_id!='{$_GET['user_id']}'";
+  $result = $db->query($query);
+  if ($result->numRows() == 0) {
+    header ("Location: adminuser.php?lastadmin={$_GET['localpart']}");
+    die;
+  }  
+
   $query = "DELETE FROM users
-    WHERE user_id={$_GET['user_id']}
-    AND domain_id={$_SESSION['domain_id']}";
+    WHERE user_id='{$_GET['user_id']}'
+    AND domain_id='{$_SESSION['domain_id']}'
+	AND (type='local' OR type='piped')";
   $result = $db->query($query);
   if (!DB::isError($result)) {
-    $query = "DELETE FROM group_contents WHERE member_id={$_GET['user_id']}";
+    $query = "DELETE FROM group_contents WHERE member_id='{$_GET['user_id']}'";
     $result = $db->query($query);
     header ("Location: adminuser.php?deleted={$_GET['localpart']}");
   } else {
@@ -22,14 +45,15 @@ if ($_GET['confirm'] == '1') {
     die;                                                      
 } else {
   $query = "SELECT user_id AS count FROM users 
-    WHERE admin=1 AND domain_id={$_SESSION['domain_id']}
-    AND user_id!={$_GET['user_id']}";
+    WHERE admin=1 AND domain_id='{$_SESSION['domain_id']}'
+	AND (type='local' OR type='piped')
+    AND user_id!='{$_GET['user_id']}'";
   $result = $db->query($query);
   if ($result->numRows() == 0) {
     header ("Location: adminuser.php?lastadmin={$_GET['localpart']}");
     die;
   }
-  $query = "SELECT localpart FROM users WHERE user_id={$_GET['user_id']}";
+  $query = "SELECT localpart FROM users WHERE user_id='{$_GET['user_id']}'";
   $result = $db->query($query);
   if ($result->numRows()) { $row = $result->fetchRow(); }
 }
